# Improve performance of rendering stdout/stderr