@jtpio 2024-12-03
@meeseeksdev please backport to 4.3.x